+++ packages/use/src/index.ts
@@ -2,3 +2,4 @@ export {
   useInputWhileMounted,
   type InputWhileMountedTargetRef,
 } from "./input-while-mounted/use-input-while-mounted.ts";
+export { useTextInput } from "./input/use-text-input.ts";